pyproject.toml
[tool.poetry]
name = "orpheus"
version = "0.1.0"
description = "The best discord bot ever made."
authors = ["Jedrzej Szadejko <jedrzej@szadejko.pl>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aiocron = "^1.6"
"discord.py" = {git = "https://github.com/Rapptz/discord.py.git", rev = "45d498c1b76deaf3b394d17ccf56112fa691d160"}
python-dotenv = "^0.19.0"
google-api-python-client = "^2.15.0"
google-auth-oauthlib = "^0.4.5"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
